Fix bug where more numbers are added to x
Result from previous calculation is stored to 'x'. If user press on
numbers right after previous calculation, those numbers are being
added to 'x' though a new calculation is intended.

Store the result in model.result to check there was a calculation
before pressing the number. If there was, clear both 'x' and 'result',
and store the number to x as a new operand.